% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baxter_metadata}
\alias{baxter_metadata}
\title{Colorectal Cancer Patients.}
\format{A data frame with 490 rows and 17 variables:
\describe{
  \item{sample}{patient ID, 7 digit number}
  \item{fit_result}{fecal immunochemical, ranges between 0 and 3000}
  \item{diagnosis_bin}{detailed cancer diagnosis, normal, high risk normal, adenoma, advanced adenoma, and cancer}
  ...
}}
\source{
\url{https://github.com/SchlossLab/Sze_CRCMetaAnalysis_mBio_2018/blob/master/data/process/baxter/metadata.tsv}
}
\usage{
baxter_metadata
}
\description{
A dataset containing the clinical information and colon cancer diagnosis for 490
patients.
}
\keyword{datasets}
