% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baxter.R
\name{get_bmi}
\alias{get_bmi}
\title{Returns an individuals body mass index, given metric units.}
\usage{
get_bmi(weight_kg, height_cm)
}
\arguments{
\item{weight_kg}{Weight in kilograms, a number.}

\item{height_cm}{Height in centimeters, a number.}
}
\value{
The BMI of a given individual, a number.
}
\description{
The squared value of weight in kilograms
divided by height in meters, which is 
converted by dividing cm by 100.
}
\examples{
get_bmi(80, 182)
get_bmi(50, 160)
}
