% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baxter.R
\name{get_bmi_category}
\alias{get_bmi_category}
\title{Determine BMI category from get_bmi function results}
\usage{
get_bmi_category(weight_kg, height_cm)
}
\arguments{
\item{weight_kg}{patient weight in kilograms}

\item{height_cm}{patient height in centimeters}
}
\value{
return BMI category for input
}
\description{
Determine BMI category from get_bmi function results
}
\examples{
get_bmi_category(100, 200)
get_bmi_category(50, 198)
get_bmi_category(180, 150)
}
