% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baxter.R
\name{is_obese}
\alias{is_obese}
\title{Determining whether a patient is obese from get_bmi_category}
\usage{
is_obese(weight_kg, height_cm)
}
\arguments{
\item{weight_kg}{patient weight in kilograms}

\item{height_cm}{patient height in centimeters}
}
\value{
True or false for obesity
}
\description{
Determining whether a patient is obese from get_bmi_category
}
\examples{
get_bmi_category(100, 200)
get_bmi_category(50, 198)
get_bmi_category(180, 150)
}
