% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalR.R
\docType{package}
\name{minimalR}
\alias{minimalR}
\alias{minimalR-package}
\title{minimalR: A package to help people learn R using microbial ecology data.}
\description{
The minimalR package contains four important functions: get_metadata, get_bmi, get_bmi_category, and is_obese.
}
\section{minimalR functions}{

The minimalR function get_metadata
}

